refacto default module layout
